Conflicts:
	build/cocos2d_libs.xcodeproj/project.pbxproj
	samples/Cpp/TestCpp/Classes/tests.h

+++ cocos/2d/platform/CCSAXParser.cpp
@@ -23,7 +23,6 @@
  ****************************************************************************/
 
 #include "CCSAXParser.h"
-#include "CCDictionary.h"
 #include "CCFileUtils.h"
 #include "tinyxml2.h"
 
@@ -102,7 +101,7 @@ bool SAXParser::init(const char *pszEncoding)
     return true;
 }
 
-bool SAXParser::parse(const char* pXMLData, unsigned int uDataLength)
+bool SAXParser::parse(const char* pXMLData, size_t uDataLength)
 {
 	tinyxml2::XMLDocument tinyDoc;
 	tinyDoc.Parse(pXMLData, uDataLength);
